Extract comparison matcher from the NumericalityMatcher module
The comparison matcher will be used now on both
ValidateNumericalityOfMatcher and on the new
ValidateComparisonOfMatcher, that's why we removed it from the
NumericalityMatcher module.
